+++ CHANGELOG.md
@@ -148,7 +148,7 @@ A new helper object is available in `utils/helpers/text`. Currently it only cont
   * has a new prop `enablePreviousButton` defaulted to `true`. Setting it to `false` will hide the previous button.
   * has a new prop `enableNextButton` defaulted to `true`. Setting it to `false` will hide the next button.
   * has a new prop `slideIndex`. Changing this prop to an index will select the corresponding slide.
-  * has a new prop `onSlideChange`. Action to be called on slide change. It will receive the slide index and the transition direction as params.
+  * has a new prop `onSlideChange`, which is an action to be called on slide change. It will receive the slide index and the transition direction as params.
   * animation between slides enhanced.
 * `Form` now has default `SaveButton` and `CancelButton` functional stateless components. The former can be overridden with a new prop of `customSaveButton`.
 * `InlineInputs` children are now wrapped by Columns by the component.